Added TypeScript to project
