+++ src/App.tsx
@@ -1,0 +1,13 @@
+import './App.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
+import { Routes, Route, BrowserRouter } from 'react-router-dom';
+import Home from "./components/public/home";
+
+function App() {
+  return (
+    <Routes>
+      <Route path="/" element={<Home/>} />
+    </Routes>
+  );
+}
+export default App;
