+++ src/components/common/jobcard.tsx
